Revise getdata.py
An revised version of the fabulous original script by jwoglom called "getdata.py". This version keeps Diabetes:m from uploading BG data back to Nightscout that was obtained from Nightscout in the first place. I also modified the original code to upload the basal insulin, proteins and fats. I can't find (for now) an "official" place for this data on Nightscout, so I have included it (with labels) in the notes section, that you can see when you hover over the dots. This version would be a complete replacement for the original "getdata.py" and would work with all of the other files jwoglom has written. No changes in his installation instructions are needed other than to replace the original version of "getdata.py" with this one and, perhaps rename this file to the original filename.