Updated version of my last getdata.py
This version corrects bugs with the prior version. All other files from the original program creator remain the same. This version writes out a string including notes, bolus, proteins and fats in the Notes section on Nightscout unless this data is empty. It also corrects the bug of some entries being multiplied by 18 and being posted to nightscout with the inflated number.